Setup and Introduction to Webpage
This is my first commit on this account. It is to be as organized and neat as possible. I wanted to get better as a coder and didn't want people to think less of me. I am going to make consistent commits on this account in order to show companies (and others interested) about my progress and the things I intend to do. This summary is a bit messy but for anyone whos reading, I hope this inspires you that no matter how bad you think you are at coding, that you can get up and practise too. Here's to doing your best in all goals you want to achieve.